requirements.txt
kafka-python
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
pandas==2.0.3
plotly==5.17.0
openai==1.3.7
python-dotenv==1.0.0
streamlit